Exercise 7.2 solution

+++ kodilla-testing/src/main/java/com/kodilla/testing/forum/statistics/ForumStatistics.java
@@ -37,7 +37,7 @@ public class ForumStatistics {
             averageCommentsByUser = commentsQuantity / usersQuantity;
         }
         if (postsQuantity > 0) {
-            averageCommentsByPost = commentsQuantity / postsQuantity;
+            averageCommentsByPost = (double) commentsQuantity / (double) postsQuantity;
         }
     }
 }
